RavenDB-18644 - address PR comments and move handler functions to processors
